[1.9.1]修复烧炼配方创建GUI经验值不接受浮点数的bug 修复Oraxen相关的各种合成bug 修复有序配方创建GUI创建配方不能正确删除空列的bug

+++ src/main/java/com/github/yufiriamazenta/craftorithm/menu/creator/CookingRecipeCreator.java
@@ -262,7 +262,7 @@ public class CookingRecipeCreator extends UnlockableRecipeCreator {
     class ExpInputPrompt implements NumberPrompt {
         @Override
         public @Nullable Prompt acceptValidatedInput(@NotNull Map<Object, Object> data, @NotNull Number number) {
-            exp = number.intValue();
+            exp = number.floatValue();
             updateExpIcon();
             openMenu();
             inConversation = false;
